+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @opencx/widget-react
 
+## 4.0.15
+
+### Patch Changes
+
+- remove dependency on `prelude` endpoint
+- Updated dependencies
+  - @opencx/widget-react-headless@4.0.15
+  - @opencx/widget-core@4.0.15
+
 ## 4.0.14
 
 ### Patch Changes